pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "athleteos"
version = "0.1.0"
description = "AthleteOS backend monorepo"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "sqlalchemy>=2.0.32",
  "alembic>=1.13.2",
  "asyncpg>=0.29.0",
  "aiosqlite>=0.20.0",
  "celery>=5.4.0",
  "redis>=5.0.8",
  "PyYAML>=6.0.2",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "slowapi>=0.1.9",
  "scikit-learn>=1.8.0",
  "imbalanced-learn>=0.14.1",
  "shap>=0.51.0",
  "email-validator>=2.2.0",
  "httpx>=0.27.0",
  "Faker>=40.12.0",
  "boto3>=1.42.82",
  "Pillow>=11.3.0",
  "PyMuPDF>=1.27.2",
  "prometheus-client>=0.21.0",
  "prometheus-fastapi-instrumentator>=7.0.0",
  "sentry-sdk[fastapi]>=2.19.0",
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "structlog>=24.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
